Paz: Use std::array where applicable
Allows modernizing the interface so raw pointers aren't being passed
around anymore.

+++ Extract/paz/ef_first.h
@@ -14,7 +14,7 @@ private:
 	void InitMountKey(CArcFile* archive) override;
 	void InitDecodeKey(CArcFile* archive) override;
 
-	u32 InitMovieTable(const u8* table) override;
+	size_t InitMovieTable(const u8* table) override;
 
 	void DecodeTable1() override;
